refactor: Added Book, BooksGrid, Bookshelf and Searchbooks ass class
Splited the provided HTML into jsx classes. Created a books class for
showing books in the panels. Created a searchcontrol to search for
books. Created a Bookshelf control for presenting books in the diffrent
shelfs and finally created a booksgrid this is used in the bookshelfs
and searchcontrol for showing a grid of books